backend/requirements.txt
Django>=4.2
djangorestframework
djongo
djangorestframework-simplejwt
pymongo
corsheaders